+++ packages/ui/certd-client/src/locales/langs/en-US/certd/navigation.ts
@@ -11,6 +11,7 @@ export default {
   siteMonitor: "Site Certificate Monitor",
   settings: "Settings",
   accessManager: "Access Management",
+  dnsPersistRecord: "DNS Persist Records",
   subDomain: "Subdomain Delegation Settings",
   pipelineGroup: "Pipeline Group Management",
   openKey: "Open API Key",