+++ packages/ui/certd-client/src/views/framework/login/index.vue
@@ -265,7 +265,6 @@ onMounted(() => {});
     .input-right {
       width: 160px;
       margin-left: 10px;
-      background: #cfcfcf !important;
     }
 
     .forge-password {